Create   Remote file download
n this code, the user is prompted to enter the URL of the remote file they want to download. Upon submitting the form, the PHP code retrieves the file URL, downloads the file content using file_get_contents(), and saves it locally using file_put_contents(). The downloaded file is saved with a unique filename to prevent conflicts.